Feat : add vfx , Refactor : hand pref, auto grab.cs
