+++ Assets/BaseballScene/Prefab/FIreEffect.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ee9cbf0e9737c1c44b875c590dd879ad
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
